refactor: split parser into smaller files
