+++ src/main.c
@@ -6,23 +6,11 @@
 /*   By: mayeung <mayeung@student.42london.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/31 16:46:40 by yublee            #+#    #+#             */
-/*   Updated: 2025/04/18 22:18:17 by mayeung          ###   ########.fr       */
+/*   Updated: 2025/04/22 16:48:30 by mayeung          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/cub3d.h"
-
-/*TODO: modify after parser*/
-
-char	*map[MAP_H + 1] = {
-	"111111",
-	"110011",
-	"100101",
-	"10P001", //(2, 3) -> (2.5, 3.5)
-	"110011",
-	"111111",
-	NULL
-};
 
 t_vecset	get_vecset(double pos_x, double pos_y, t_cardinal direction)
 {
@@ -62,30 +50,10 @@ t_vecset	get_vecset(double pos_x, double pos_y, t_cardinal direction)
 	return (vecset);
 }
 
-static t_colorset	get_colorset(int color1, int color2)
-{
-	t_colorset	colorset;
-
-	colorset.color_ceiling = color1;
-	colorset.color_floor = color2;
-	return (colorset);
-}
-
-/***************************/
-
 int	main(int argc, char **argv)
 {
 	t_vars		vars;
 	t_vecset	vecset;
-
-	/*TODO: get from parser*/
-	// vars.map = map;
-	// vars.map_height = MAP_H;
-	// vars.map_width = MAP_W;
-	// vecset = get_vecset(2.5, 3.5, NORTH);
-	// vars.vecset = &vecset;
-	// vars.colorset = get_colorset((int)0xffe6dcd1, (int)0xff2b1904);
-	/********************* */
 
 	ft_bzero(&vars, sizeof(t_vars));
 	if (argc != 2)
@@ -99,7 +67,6 @@ int	main(int argc, char **argv)
 	draw_image(&vars);
 	mlx_put_image_to_window(vars.mlx, vars.win, vars.img.img_ptr, 0, 0);
 	mlx_hook(vars.win, 2, 1L << 0, handle_key_input, &vars);
-	// mlx_key_hook(vars.win, handle_key_input, &vars);
 	mlx_hook(vars.win, 17, 1L << 2, close_game, &vars);
 	mlx_loop(vars.mlx);
 	return (0);
